requirements.txt
asgiref==3.11.0
colorama==0.4.6
decorator==5.2.1
Django==5.2.11
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.4.2
pillow==11.3.0
proglog==0.1.12
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
tqdm==4.67.3
tzdata==2025.3
django-filter==24.3
ruff==0.6.9
